requirements.txt
uniswap-python
ccxt